requirements.txt
aio-pika==6.8.0
aiormq==3.3.1
idna==3.1
multidict==5.1.0
numpy==1.20.2
pamqp==2.3.0
pkg-resources==0.0.0
pygame==2.0.1
PyYAML==5.4.1
yarl==1.6.3
